requirements.txt
notion-client==2.2.1
feedparser==6.0.11
backoff==2.2.1
python-dateutil==2.9.0.post0
